# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "opendataAPI",
        "description": "Open data API from open dataset",
        "contact": {
            "email": "mellyniaramadhan762@gmail.com"
        },
        "license": {
            "name": "mellyniaramadhan",
            "url": "http://your-license-url.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/OpenDataJabar": {
            "post": {
                "tags": [
                    "Consumsion Taxs"
                ],
                "summary": "Create a new Consumsion Taxs record.",
                "operationId": "46b783713c91a419e4ff3810be94b311",
                "requestBody": {
                    "description": "Data to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "CodeProvince": {
                                        "type": "integer"
                                    },
                                    "ProvinceName": {
                                        "type": "string"
                                    },
                                    "CodeRegencyCity": {
                                        "type": "integer"
                                    },
                                    "RegencyNameCity": {
                                        "type": "string"
                                    },
                                    "NumberScorePPH": {
                                        "type": "integer"
                                    },
                                    "Unit": {
                                        "type": "string"
                                    },
                                    "Year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "CodeProvince": "example CodeProvince",
                                "ProvinceName": "example ProvinceName",
                                "CodeRegencyName": "example CodeRegencyName",
                                "RegencyNameCity": "Example RegencyNameCity",
                                "NumberScorePPH": "Example NumberScorePPH",
                                "Unit": "Example Unit",
                                "Year": "Example Year"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "CodeProvince": {
                                            "type": "integer",
                                            "example": "32"
                                        },
                                        "ProvinceName": {
                                            "type": "string",
                                            "example": "Jawa Barat"
                                        },
                                        "CodeRegencyName": {
                                            "type": "integer",
                                            "example": 3201
                                        },
                                        "RegencyNameCity": {
                                            "type": "string",
                                            "example": "Kabupaten Bogor"
                                        },
                                        "NumberScorePPH": {
                                            "type": "integer",
                                            "example": 81
                                        },
                                        "Unit": {
                                            "type": "string",
                                            "example": "POINT"
                                        },
                                        "Year": {
                                            "type": "integer",
                                            "example": 2021
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "fieldName": [
                                                    "Error message"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/OpenDataJabar/MangoProductions": {
            "get": {
                "tags": [
                    "Mango Productions"
                ],
                "summary": "Get a paginated list of Mango Productions.",
                "operationId": "4a4bf22e3a70994c6e6235905762181f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MangoProductions"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/OpenDatajabar": {
            "post": {
                "tags": [
                    "Mango Productions"
                ],
                "summary": "Create a new Mango Productions record.",
                "operationId": "5860c4c9147513ba89124acbe2be2d27",
                "requestBody": {
                    "description": "Data to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "CodeProvince": {
                                        "type": "integer"
                                    },
                                    "ProvinceName": {
                                        "type": "string"
                                    },
                                    "CodeRegencyCity": {
                                        "type": "integer"
                                    },
                                    "RegencyNameCity": {
                                        "type": "string"
                                    },
                                    "MangoProductions": {
                                        "type": "integer"
                                    },
                                    "Unit": {
                                        "type": "string"
                                    },
                                    "Year": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "CodeProvince": "example CodeProvince",
                                "ProvinceName": "example ProvinceName",
                                "CodeRegencyCity": "example CodeRegencyCity",
                                "RegencyCityName": "Example RegencyCityName",
                                "MangoProductions": "Example MangoProductions",
                                "Unit": "Example Unit",
                                "Year": "Example Year"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "CodeProvince": {
                                            "type": "integer",
                                            "example": "32"
                                        },
                                        "ProvinceName": {
                                            "type": "string",
                                            "example": "Jawa Barat"
                                        },
                                        "CodeRegencyCity": {
                                            "type": "integer",
                                            "example": 3201
                                        },
                                        "RegencyCityName": {
                                            "type": "string",
                                            "example": "Kabupaten Bogor"
                                        },
                                        "MangoProductions": {
                                            "type": "integer",
                                            "example": 29405
                                        },
                                        "Unit": {
                                            "type": "string",
                                            "example": "KUINTAL"
                                        },
                                        "Year": {
                                            "type": "string",
                                            "example": "2021"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "fieldName": [
                                                    "Error message"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MangoProductions": {
                "properties": {
                    "CodeProvince": {
                        "type": "integer"
                    },
                    "ProvinceName": {
                        "type": "string"
                    },
                    "CodeRegencyCity": {
                        "type": "integer"
                    },
                    "RegencyNameCity": {
                        "type": "string"
                    },
                    "MangoProductions": {
                        "type": "string"
                    },
                    "Unit": {
                        "type": "string"
                    },
                    "Year": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}